Scrutinizer Recommendation
Mktime uses int arguments rather than string.

+++ src/PhpSpreadsheet/Shared/OLE.php
@@ -505,7 +505,7 @@ class OLE
         // days from 1-1-1601 until the beggining of UNIX era
         $days = 134774;
         // calculate seconds
-        $big_date = $days * 24 * 3600 + mktime(date('H', $date), date('i', $date), date('s', $date), date('m', $date), date('d', $date), date('Y', $date));
+        $big_date = $days * 24 * 3600 + mktime((int) date('H', $date), (int) date('i', $date), (int) date('s', $date), (int) date('m', $date), (int) date('d', $date), (int) date('Y', $date));
         // multiply just to make MS happy
         $big_date *= 10000000;
